Refactors web element as JSON wire spec.
